+++ R/addin.R
@@ -17,13 +17,13 @@
 #'   or in a web browser (`"browser"`).
 #'
 #' @export
-#' @examples
-#' \dontshow{ asNamespace("pkgsearch")$exif(FALSE, \{ }
+#' @section Examples:
+#' ```
 #' pkg_search_addin()
 #'
 #' # Start with a search query
 #' pkg_search_addin("permutation test")
-#' \dontshow{ \}) }
+#' ```
 
 pkg_search_addin <- function(
   query = "",
@@ -82,10 +82,18 @@ pkg_search_addin <- function(
   }
 
   searchResults <- function(id) {
-    shiny::htmlOutput(paste0("results-", id))
+    shiny::tags$div(
+      shiny::tags$div(
+        id = paste0("spin-", id),
+        class = "spinner",
+        shiny::HTML(spin_html())
+      ),
+      shiny::htmlOutput(paste0("results-", id))
+    )
   }
 
   ui <- shiny::navbarPage(
+    windowTitle = "Search CRAN packages",
     title = shiny::div(
       shiny::div(
         id = "rhub-logo",
@@ -96,6 +104,10 @@ pkg_search_addin <- function(
           ),
           href = "https://docs.r-hub.io"
         )
+        ),
+      shiny::div(
+        id = "done-button",
+        shiny::actionButton("done", label = "Done", class = "btn btn-primary")
       )
     ),
     shiny::tabPanel("Search", searchQuery("search"), searchResults("search")),
@@ -109,6 +121,7 @@ pkg_search_addin <- function(
     header = shiny::tagList(
       shiny::tags$head(
         shiny::tags$style(addin_styles()),
+        shiny::tags$style(spin_css()),
         shiny::tags$link(
           rel = "stylesheet",
           type = "text/css",
@@ -141,59 +154,76 @@ pkg_search_addin <- function(
 
   server <- function(input, output, session) {
     session$sendCustomMessage("selectSearch", "select")
+
+    shiny::observeEvent(input$done, shiny::stopApp())
+
     output$`results-search` <- shiny::renderUI({
+      if (input$`query-search` != "") {
+        shinyjs::runjs("$(\"#spin-search\").css(\"display\",\"block\");")
+      }
       ret <- simple_search(
         input$`query-search`,
         reactives$`search-prev`,
         reactives$`search-next`
       )
       shinyjs::runjs("window.scrollTo(0, 0)")
+      shinyjs::runjs("$(\"#spin-search\").css(\"display\",\"none\");")
       ret
     })
 
     output$`results-new` <- shiny::renderUI({
+      shinyjs::runjs("$(\"#spin-new\").css(\"display\",\"block\");")
       ret <- new_search(
         reactives$`new-prev`,
         reactives$`new-next`
       )
       shinyjs::runjs("window.scrollTo(0, 0)")
+      shinyjs::runjs("$(\"#spin-new\").css(\"display\",\"none\");")
       ret
     })
 
     output$`results-topdl` <- shiny::renderUI({
+      shinyjs::runjs("$(\"#spin-topdl\").css(\"display\",\"block\");")
       ret <- topdl_search(
         reactives$`topdl-prev`,
         reactives$`topdl-next`
       )
       shinyjs::runjs("window.scrollTo(0, 0)")
+      shinyjs::runjs("$(\"#spin-topdl\").css(\"display\",\"none\");")
       ret
     })
 
     output$`results-topdep` <- shiny::renderUI({
+      shinyjs::runjs("$(\"#spin-topdep\").css(\"display\",\"block\");")
       ret <- topdep_search(
         reactives$`topdep-prev`,
         reactives$`topdep-next`
       )
       shinyjs::runjs("window.scrollTo(0, 0)")
+      shinyjs::runjs("$(\"#spin-topdep\").css(\"display\",\"none\");")
       ret
     })
 
     output$`results-trend` <- shiny::renderUI({
+      shinyjs::runjs("$(\"#spin-trend\").css(\"display\",\"block\");")
       ret <- trend_search(
         reactives$`trend-prev`,
         reactives$`trend-next`
       )
       shinyjs::runjs("window.scrollTo(0, 0)")
+      shinyjs::runjs("$(\"#spin-trend\").css(\"display\",\"none\");")
       ret
     })
 
     output$`results-maint` <- shiny::renderUI({
+      shinyjs::runjs("$(\"#spin-maint\").css(\"display\",\"block\");")
       ret <- maint_search(
         input$`query-maint`,
         reactives$`maint-prev`,
         reactives$`maint-next`
       )
       shinyjs::runjs("window.scrollTo(0, 0)")
+      shinyjs::runjs("$(\"#spin-maint\").css(\"display\",\"none\");")
       ret
     })
 
@@ -353,7 +383,7 @@ pkg_search_addin <- function(
   }
 
   dep_list0 <- function() {
-    deps <- query("/-/deps/devel")
+    deps <- crandb_query("/-/deps/devel")
     tibble::tibble(
       package = names(deps),
       count = unlist(deps)
@@ -436,7 +466,7 @@ pkg_search_addin <- function(
     keys <- map_chr(keys, utils::URLencode)
     keys <- paste0('"', keys, '"', collapse = ",")
     url <- paste0(ep, "?keys=[", keys, "]")
-    ret <- query(url)
+    ret <- crandb_query(url)
     tibble::tibble(email = ret[,1], package = ret[,2])
   }
 
@@ -483,6 +513,7 @@ pkg_search_addin <- function(
     utils::browseURL(url)
   }
 
+  on.exit(tryCatch(shiny::stopApp(), error = function(e) NULL), add = TRUE)
   shiny::runGadget(ui, server, viewer = viewer)
 }
 
@@ -518,6 +549,7 @@ addin_styles <- function() {
           .btn-package {
             font-size: 140%;
             color: rgb(51, 122, 183);
+            background-color: #fff;
             border: none;
             padding-left: 0px;
           }
@@ -542,8 +574,79 @@ addin_styles <- function() {
             top: 0px;
             margin-top: -10px;
           }
+          #done-button {
+            position: absolute;
+            top: 10px;
+            right: 10px;
+          }
+          .spinner {
+            display: none;
+          }
           "
   )
+}
+
+spin_css <- function() {
+  ".lds-ellipsis {
+    display: inline-block;
+    position: relative;
+    width: 40px;
+    height: 40px;
+  }
+  .lds-ellipsis div {
+    position: absolute;
+    top: 13px;
+    width: 13px;
+    height: 13px;
+    border-radius: 50%;
+    background: #569fde;
+    animation-timing-function: cubic-bezier(0, 1, 1, 0);
+  }
+  .lds-ellipsis div:nth-child(1) {
+    left: 8px;
+    animation: lds-ellipsis1 0.6s infinite;
+  }
+  .lds-ellipsis div:nth-child(2) {
+    left: 8px;
+    animation: lds-ellipsis2 0.6s infinite;
+  }
+  .lds-ellipsis div:nth-child(3) {
+    left: 32px;
+    animation: lds-ellipsis2 0.6s infinite;
+  }
+  .lds-ellipsis div:nth-child(4) {
+    left: 56px;
+    animation: lds-ellipsis3 0.6s infinite;
+  }
+  @keyframes lds-ellipsis1 {
+    0% {
+      transform: scale(0);
+    }
+    100% {
+      transform: scale(1);
+    }
+  }
+  @keyframes lds-ellipsis3 {
+    0% {
+      transform: scale(1);
+    }
+    100% {
+      transform: scale(0);
+    }
+  }
+  @keyframes lds-ellipsis2 {
+    0% {
+      transform: translate(0, 0);
+    }
+    100% {
+      transform: translate(24px, 0);
+    }
+  }
+  "
+}
+
+spin_html <- function() {
+  "<div class=\"lds-ellipsis\"><div></div><div></div><div></div><div></div></div>"
 }
 
 format_addin_results <- function(results, id) {